% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_vline.R
\name{add_vline}
\alias{add_vline}
\title{Add a vertical line to ggplot}
\usage{
add_vline(p, ltext, ldate, linetype = "solid")
}
\arguments{
\item{p}{The plot}

\item{ltext}{Text to add parallel to the line}

\item{ldate}{Date on which to put the line}
}
\description{
Adds a line with annotation at a specified date (x-axis)
}
