% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdataitaly.R
\name{getdataitaly}
\alias{getdataitaly}
\title{Download confirmed, deaths and recovered from Italian CDC}
\usage{
getdataitaly()
}
\value{
Saves a data frame with region, date, positive, death, recovered, total hospitalized, and ICU to the data folder.
}
\description{
Date is posted on the github site daily at about 3pm EST. github/pcm-dpc/COVID-19
}
\examples{
\dontrun{
getdataitaly()
}
}
