% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdatawa.R
\name{getdatawa}
\alias{getdatawa}
\title{Get WA data from file downloaded from WA DOH}
\usage{
getdatawa()
}
\value{
Returns a data frame that is used in `getdatastates()`
}
\description{
Need to manually download file. Save to .csv. Save tab for total and tab for deaths separately. `PUBLIC-CDC-Event-Date-SARS.csv` and `PUBLIC-CDC-Event-Date-SARS-deaths.csv`
}
\examples{
\dontrun{
getdatawa()
}
}
