% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdataworld.R
\name{getdataworld}
\alias{getdataworld}
\title{Download confirmed, deaths and recovered from JHU database}
\usage{
getdataworld()
}
\value{
Saves a data frame with date, region, positive, death, and recovered to the data folder.
}
\description{
Download confirmed, deaths and recovered from JHU database
}
\examples{
\dontrun{
getdataworld()
}
}
