% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdataworld_oldjhu.R
\name{getdataworld_oldjhu}
\alias{getdataworld_oldjhu}
\title{Download confirmed, deaths and recovered from JHU database}
\usage{
getdataworld_oldjhu()
}
\value{
Saves a data frame with date, region, positive, death, and recovered to the data folder.
}
\description{
This is the old JHU data before they changed to new data format and stopped recording recovered.
}
\examples{
\dontrun{
getdataworld_oldjhu()
}
}
