% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{regions}
\alias{regions}
\title{Show the possible regions in the data files}
\usage{
regions()
}
\description{
Type regions() and then it will ask what data set you want to use. You can also enter
a pattern to reduce the regions. For example, type "WA US" to show all regions with
the text "WA US". Use "China" to show all regions in China.
}
\examples{
\dontrun{
regions()
Which dataset? a) states, b) world, c) italy? (a/b/c) b
Search pattern? (return to show all) WA US
}
}
