% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{states}
\alias{states}
\title{US states coronavirus data}
\format{
A data frame with 7 variables:
\describe{
  \item{date}{date. Data are for the previous day.}
  \item{region}{State name}
  \item{positive}{Cumulative positives.}
  \item{negative}{Cumulative negatives (from testing).}
  \item{hospitalized}{Current number in the hospital. aka totale_ospedalizzati}
  \item{death}{Cumulative deaths}
  \item{total tests}{Cumulative tests. These data are spotty and perhaps not very accurate.}
}
}
\usage{
states
}
\description{
State data from Covid Tracking. The data cutoff is 4pm EST. 
So some states are delayed by one day because they report after that time. Data are 
downloaded from https://covidtracking.com/api/states/daily
See `getdatastates.R` for processing of the data file.
}
\keyword{datasets}
