% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdataworld.R
\name{getdataworld}
\alias{getdataworld}
\title{Download confirmed and deaths from new JHU database}
\usage{
getdataworld()
}
\value{
Saves a data frame with date, region, positive, and death to the data folder.
}
\description{
Download confirmed and deaths from new JHU database
}
\examples{
\dontrun{
getdataworld()
}
}
